+++ packages/katana_theme/lib/katana_theme.dart
@@ -10,6 +10,7 @@ library;
 
 // Dart imports:
 import "dart:async";
+import "dart:math" as math;
 import "dart:ui" as ui;
 import "dart:ui";